requirements.txt
agno
pyside6
httpx
python-dotenv
groq
markdown
